Hopefully work around deadlock when Consul's being a bitch
NEEDS-REFACTORING ALERT

Various circumstances seem to cause Consul to wig out and return 500s
when polling for health checks, such as the cluster not having a
leader.  This also causes the session to be invalidated, for obvious
reasons.  When this happened, the mainLoop would attempt to abort the
health checker when it was no longer select'ing the done channel.
This change ensures that the healthResultsAbort channel is closed and
set to nil when the health checker fails.